multi-flask/flask-app/requirements.txt
Flask
mysqlclient
mysql-connector-python